third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/KeyframeEffect/constructor-expected.txt

This is a testharness.js-based test.
Found 121 tests; 66 PASS, 55 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS a KeyframeEffectReadOnly can be constructed with no frames
PASS easing values are parsed correctly when passed to the KeyframeEffectReadOnly constructor in KeyframeEffectOptions
PASS invalid easing values are correctly rejected when passed to the KeyframeEffectReadOnly constructor in KeyframeEffectOptions
FAIL composite values are parsed correctly when passed to the KeyframeEffectReadOnly constructor in property-indexed keyframes assert_equals: resulting composite for 'accumulate' expected "accumulate" but got "replace"
FAIL composite values are parsed correctly when passed to the KeyframeEffectReadOnly constructor in regular keyframes assert_equals: resulting composite for 'accumulate' expected "accumulate" but got "replace"
FAIL composite value is absent if the composite operation specified on the keyframe effect is being used assert_equals: resulting composite for 'replace' expected (undefined) undefined but got (string) "replace"
FAIL a KeyframeEffectReadOnly can be constructed with a one property two value property-indexed keyframes specification assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a one property two value property-indexed keyframes specification roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one shorthand property two value property-indexed keyframes specification assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,offset" but got "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
PASS a KeyframeEffectReadOnly constructed with a one shorthand property two value property-indexed keyframes specification roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a two property (one shorthand and one of its longhand components) two value property-indexed keyframes specification assert_equals: number of frames expected 2 but got 4
PASS a KeyframeEffectReadOnly constructed with a two property (one shorthand and one of its longhand components) two value property-indexed keyframes specification roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a two property two value property-indexed keyframes specification assert_equals: number of frames expected 2 but got 4
PASS a KeyframeEffectReadOnly constructed with a two property two value property-indexed keyframes specification roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a two property property-indexed keyframes specification with different numbers of values assert_equals: number of frames expected 3 but got 5
PASS a KeyframeEffectReadOnly constructed with a two property property-indexed keyframes specification with different numbers of values roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a property-indexed keyframes specification with an invalid value assert_equals: number of frames expected 5 but got 10
PASS a KeyframeEffectReadOnly constructed with a property-indexed keyframes specification with an invalid value roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one property two value property-indexed keyframes specification that needs to stringify its values assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,offset,opacity" but got "composite,computedOffset,easing,offset,opacity"
PASS a KeyframeEffectReadOnly constructed with a one property two value property-indexed keyframes specification that needs to stringify its values roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a property-indexed keyframes specification with a CSS variable reference assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a property-indexed keyframes specification with a CSS variable reference roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a property-indexed keyframes specification with a CSS variable reference in a shorthand property assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,offset" but got "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
FAIL a KeyframeEffectReadOnly constructed with a property-indexed keyframes specification with a CSS variable reference in a shorthand property roundtrips assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset" but got "composite,computedOffset,easing,offset"
FAIL a KeyframeEffectReadOnly can be constructed with a one property one value property-indexed keyframes specification assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a one property one value property-indexed keyframes specification roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one property one non-array value property-indexed keyframes specification assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a one property one non-array value property-indexed keyframes specification roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one property two value property-indexed keyframes specification where the first value is invalid assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
PASS a KeyframeEffectReadOnly constructed with a one property two value property-indexed keyframes specification where the first value is invalid roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one property two value property-indexed keyframes specification where the second value is invalid assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a one property two value property-indexed keyframes specification where the second value is invalid roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one property one keyframe sequence assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a one property one keyframe sequence roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one property two keyframe sequence assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a one property two keyframe sequence roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a two property two keyframe sequence assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset,top" but got "composite,computedOffset,easing,left,offset,top"
PASS a KeyframeEffectReadOnly constructed with a two property two keyframe sequence roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one shorthand property two keyframe sequence assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,offset" but got "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
PASS a KeyframeEffectReadOnly constructed with a one shorthand property two keyframe sequence roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a two property (a shorthand and one of its component longhands) two keyframe sequence assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,marginTop,offset" but got "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
PASS a KeyframeEffectReadOnly constructed with a two property (a shorthand and one of its component longhands) two keyframe sequence roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence with duplicate values for a given interior offset assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence with duplicate values for a given interior offset roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence with duplicate values for offsets 0 and 1 assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence with duplicate values for offsets 0 and 1 roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a two property four keyframe sequence assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a two property four keyframe sequence roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a single keyframe sequence with omitted offset assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
FAIL a KeyframeEffectReadOnly constructed with a single keyframe sequence with omitted offset roundtrips Failed to construct 'KeyframeEffectReadOnly': Non numeric offset provided
FAIL a KeyframeEffectReadOnly can be constructed with a single keyframe sequence with null offset assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
FAIL a KeyframeEffectReadOnly constructed with a single keyframe sequence with null offset roundtrips Failed to construct 'KeyframeEffectReadOnly': Non numeric offset provided
FAIL a KeyframeEffectReadOnly can be constructed with a single keyframe sequence with string offset assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a single keyframe sequence with string offset roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one property keyframe sequence with some omitted offsets assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
FAIL a KeyframeEffectReadOnly constructed with a one property keyframe sequence with some omitted offsets roundtrips Failed to construct 'KeyframeEffectReadOnly': Non numeric offset provided
FAIL a KeyframeEffectReadOnly can be constructed with a one property keyframe sequence with some null offsets assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
FAIL a KeyframeEffectReadOnly constructed with a one property keyframe sequence with some null offsets roundtrips Failed to construct 'KeyframeEffectReadOnly': Non numeric offset provided
FAIL a KeyframeEffectReadOnly can be constructed with a two property keyframe sequence with some omitted offsets assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset,top" but got "composite,computedOffset,easing,left,offset,top"
FAIL a KeyframeEffectReadOnly constructed with a two property keyframe sequence with some omitted offsets roundtrips Failed to construct 'KeyframeEffectReadOnly': Non numeric offset provided
FAIL a KeyframeEffectReadOnly can be constructed with a one property keyframe sequence with all omitted offsets assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
FAIL a KeyframeEffectReadOnly constructed with a one property keyframe sequence with all omitted offsets roundtrips Failed to construct 'KeyframeEffectReadOnly': Non numeric offset provided
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence with different easing values, but the same easing value for a given offset assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence with different easing values, but the same easing value for a given offset roundtrips
PASS a KeyframeEffectReadOnly can be constructed with a keyframe sequence with different composite values, but the same composite value for a given offset
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence with different composite values, but the same composite value for a given offset roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a one property two keyframe sequence that needs to stringify its values assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,offset,opacity" but got "composite,computedOffset,easing,offset,opacity"
PASS a KeyframeEffectReadOnly constructed with a one property two keyframe sequence that needs to stringify its values roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence with a CSS variable reference assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
FAIL a KeyframeEffectReadOnly constructed with a keyframe sequence with a CSS variable reference roundtrips Failed to construct 'KeyframeEffectReadOnly': Non numeric offset provided
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence with a CSS variable reference in a shorthand property assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,offset" but got "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
FAIL a KeyframeEffectReadOnly constructed with a keyframe sequence with a CSS variable reference in a shorthand property roundtrips Failed to construct 'KeyframeEffectReadOnly': Non numeric offset provided
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence where shorthand precedes longhand assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,marginRight,offset" but got "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence where shorthand precedes longhand roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence where longhand precedes shorthand assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,marginRight,offset" but got "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence where longhand precedes shorthand roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence where lesser shorthand precedes greater shorthand assert_equals: properties on ComputedKeyframe #0 expected "border,borderLeft,computedOffset,easing,offset" but got "borderBottomColor,borderBottomStyle,borderBottomWidth,borderImageOutset,borderImageRepeat,borderImageSlice,borderImageSource,borderImageWidth,borderLeftColor,borderLeftStyle,borderLeftWidth,borderRightColor,borderRightStyle,borderRightWidth,borderTopColor,borderTopStyle,borderTopWidth,composite,computedOffset,easing,offset"
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence where lesser shorthand precedes greater shorthand roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence where greater shorthand precedes lesser shorthand assert_equals: properties on ComputedKeyframe #0 expected "border,borderLeft,computedOffset,easing,offset" but got "borderBottomColor,borderBottomStyle,borderBottomWidth,borderImageOutset,borderImageRepeat,borderImageSlice,borderImageSource,borderImageWidth,borderLeftColor,borderLeftStyle,borderLeftWidth,borderRightColor,borderRightStyle,borderRightWidth,borderTopColor,borderTopStyle,borderTopWidth,composite,computedOffset,easing,offset"
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence where greater shorthand precedes lesser shorthand roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a two property keyframe sequence where one property is missing from the first keyframe assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a two property keyframe sequence where one property is missing from the first keyframe roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a two property keyframe sequence where one property is missing from the last keyframe assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset,top" but got "composite,computedOffset,easing,left,offset,top"
PASS a KeyframeEffectReadOnly constructed with a two property keyframe sequence where one property is missing from the last keyframe roundtrips
FAIL a KeyframeEffectReadOnly can be constructed with a keyframe sequence with repeated values at offset 1 with different easings assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,left,offset" but got "composite,computedOffset,easing,left,offset"
PASS a KeyframeEffectReadOnly constructed with a keyframe sequence with repeated values at offset 1 with different easings roundtrips
PASS KeyframeEffectReadOnly constructor throws with keyframes with an out-of-bounded positive offset
PASS KeyframeEffectReadOnly constructor throws with keyframes with an out-of-bounded negative offset
PASS KeyframeEffectReadOnly constructor throws with keyframes not loosely sorted by offset
PASS KeyframeEffectReadOnly constructor throws with property-indexed keyframes with an invalid easing value
PASS KeyframeEffectReadOnly constructor throws with a keyframe sequence with an invalid easing value
FAIL KeyframeEffectReadOnly constructor throws with keyframes with an invalid composite value assert_throws: function "function () {
      new KeyframeEffectReadOnly(target, subtest.input);
    }" did not throw
FAIL a KeyframeEffectReadOnly constructed without any KeyframeEffectOptions object assert_equals: default composite expected (string) "replace" but got (undefined) undefined
PASS a KeyframeEffectReadOnly constructed by an empty KeyframeEffectOptions object
PASS a KeyframeEffectReadOnly constructed by a normal KeyframeEffectOptions object
PASS a KeyframeEffectReadOnly constructed by a double value
PASS a KeyframeEffectReadOnly constructed by +Infinity
PASS a KeyframeEffectReadOnly constructed by an Infinity duration
PASS a KeyframeEffectReadOnly constructed by an auto duration
PASS a KeyframeEffectReadOnly constructed by an Infinity iterations
PASS a KeyframeEffectReadOnly constructed by an auto fill
PASS a KeyframeEffectReadOnly constructed by a forwards fill
PASS Invalid KeyframeEffectReadOnly option by -Infinity
PASS Invalid KeyframeEffectReadOnly option by NaN
PASS Invalid KeyframeEffectReadOnly option by a negative value
PASS Invalid KeyframeEffectReadOnly option by a negative Infinity duration
PASS Invalid KeyframeEffectReadOnly option by a NaN duration
PASS Invalid KeyframeEffectReadOnly option by a negative duration
PASS Invalid KeyframeEffectReadOnly option by a string duration
PASS Invalid KeyframeEffectReadOnly option by a negative Infinity iterations
PASS Invalid KeyframeEffectReadOnly option by a NaN iterations
PASS Invalid KeyframeEffectReadOnly option by a negative iterations
PASS Invalid KeyframeEffectReadOnly option by a blank easing
PASS Invalid KeyframeEffectReadOnly option by an unrecognized easing
PASS Invalid KeyframeEffectReadOnly option by an 'initial' easing
PASS Invalid KeyframeEffectReadOnly option by an 'inherit' easing
PASS Invalid KeyframeEffectReadOnly option by a variable easing
PASS Invalid KeyframeEffectReadOnly option by a multi-value easing
FAIL a KeyframeEffectReadOnly constructed with null target assert_equals: Effect created with null target has correct target expected (object) null but got (undefined) undefined
PASS KeyframeEffect constructor creates an AnimationEffectTiming timing object
FAIL KeyframeEffect constructor propagates exceptions generated by accessing the options object assert_throws: function "function () {
    new KeyframeEffect(target, { get left() { throw test_error }})
  }" did not throw
Harness: the test ran to completion.
